fix: import conflicts in customElementRegistry
