+++ packages/all/src/wired-elements.ts
@@ -7,7 +7,6 @@ export { WiredDialog } from 'wired-dialog';
 export { WiredDivider } from 'wired-divider';
 export { WiredFab } from 'wired-fab';
 export { WiredIcon } from 'wired-icon';
-export { iconsetLoader } from 'wired-icon/lib/iconset'
 export { WiredIconButton } from 'wired-icon-button';
 export { WiredImage } from 'wired-image';
 export { WiredInput } from 'wired-input';
